Implement q_reverse_K
Because if the number of nodes is not a multiple of k, the left-out
nodes would remain as it is. I count how many couples of k the queue has
first, and then use a flag variable 'cnt_n' to record the current round.
In every round, I move the current node to the first position in that
round.